backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
supabase==2.3.4
httpx>=0.24,<0.26
PyMuPDF==1.23.22
python-jose[cryptography]==3.3.0
tiktoken==0.6.0
numpy==1.26.4
python-dotenv==1.0.1
